requirements/dev.txt
-r common.txt
coverage
sphinx
pylint
pep8
django-debug-toolbar==1.8
django-extensions
tox